Today's Odds

Determine today's games' odds (if today has games), or a specified date's odds

todayOdds(
  params = NULL,
  today = Sys.Date(),
  schedule = HockeyModel::schedule,
  expected_mean = NULL,
  season_percent = NULL,
  include_xG = FALSE
)

Arguments

params
The named list containing m, rho, beta, eta, and k. See updateDC for information on the params list

today
The date for which you want game odds

schedule
The schedule, default to internal schedule

expected_mean
the mean lambda & mu, used only for regression

season_percent
the percent complete of the season, used for regression

include_xG
Whether to include xG values in reported odds

Value

a data frame of HomeTeam, AwayTeam, HomeWin, AwayWin, Draw, or NULL if no games today

Examples

todayOdds(today=as.Date("2019-11-01"))

#>              HomeTeam              AwayTeam   HomeWin   AwayWin      Draw
#> 1   New Jersey Devils   Philadelphia Flyers 0.3636543 0.4113869 0.2249587
#> 2  New York Islanders   Tampa Bay Lightning 0.3816925 0.3528302 0.2654773
#> 3 Washington Capitals        Buffalo Sabres 0.6043217 0.1950323 0.2006460
#> 4 Carolina Hurricanes     Detroit Red Wings 0.6599250 0.1461731 0.1939020
#> 5     St. Louis Blues Columbus Blue Jackets 0.5044862 0.2683713 0.2271425
#> 6  Colorado Avalanche          Dallas Stars 0.4779002 0.2775716 0.2445282
#> 7       Anaheim Ducks     Vancouver Canucks 0.3530553 0.3932677 0.2536770
#> 8     San Jose Sharks         Winnipeg Jets 0.3499994 0.4174217 0.2325789
#>       GameID
#> 1 2019020196
#> 2 2019020197
#> 3 2019020198
#> 4 2019020199
#> 5 2019020200
#> 6 2019020201
#> 7 2019020202
#> 8 2019020203